バックボタン処理の改善とスクリーンシェアが切断されないバグを修正
